Optimizing multi threaded execution.
We have a set of started threads. If urlQueue is empty and if any thread
is running then that thread might add urls to urlQueue, so we keep
waiting for that running thread to join back.

Previously it will wait for every started thread, now it will wait only
if urlQueue is empty.